Rename component files to Pascal Case
